Fix link error detection in glsl-link-initializer-03, and make more robust
Previously, this test was attempting to check the link status of a
program linked by piglit_link_simple_program() using glGetProgramiv().
But piglit_link_simple_program() returns 0 on a link failure.  So the
test really just needs to verify that the two program handles returned
are nonzero.

Also, the test used to rely on assumptions about the order in which
shaders are attached.  Now it tries both possible attachment orders.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>