gl-3.0: Verify linking errors to due invalid glBindFragDataLocation assignments
This test fails on AMD's closed-source driver and NVIDIA's
closed-source driver because they both allow an array to be assigned
to location GL_MAX_DRAW_BUFFERS-1.  See comments in the test for the
reasons this should be rejected.  AMD's closed-source driver also
fails to reject overlapping location assignments for 'a[1]' and 'v'.

v2: Verify that setting a and v to valid locations after a failing
link can link successfully.  Suggested by Paul Berry.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>