egl: Add all EGL tests to all_egl.tests
The EGL tests were built by CMake, but since they were not enumerated in
any Python file one had to run them manually. This meant the EGL tests
were never ran.

We can't add the EGL tests to all.tests because Windows doesn't support
EGL. Ideally, we could add the EGL tests to all.tests and the test runner
would skip the tests if libEGL was not present. However, the test runner
is not too bright. This patch lists the tests in a separate file as
a workaround until we find a better solution.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>