Fix redundant static-write-varying-01.frag tests.
When the static-write-varying-01.frag tests were introduced, they were
accidentally made the same as the corresponding
static-write-varying-02.frag.

The original intent (from talking to Chad) was for one set of tests
verify that varyings can't be written to via direct assignment, and
the other set of tests to verify that varyings can't be written to via
passing to a function as an "out" parameter.

This patch updates static-write-varying-01.frag to reflect the
original intent.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>