glsl-1.30: Verify that gl_Frag{Color,Data} and a function out can be written
These tests try to trick the GLSL compiler by writing a built-in
fragment shader output and a function "out" variable.  The compiler is
supposed to generate an error if a built-in output and a user-defined
shader "out" variable are written.  Writing to a built-in and a
funciton "out" should be allowed.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>