fix egl-query-surface argv parsing
This fixes:
1. If you passed any option not recognized by egl-query-surface,
but that *is* recognized by egl_util_run() like '-auto', it would
usually fail with usage().
2. If it didn't fail, it was because the loop skips over the next
argument after it finds one it recognizes.
3. If a recognized argument is the last/only argument in argv,
egl_util_run() would segfault with a null deref in strcmp() because
argc would be off by one, because of problem 2.

The only way the test didn't crash or fail was if two arguments
were passed in this exact order:

bin/egl-query-surface --bad-surface -auto

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>