arb_blend_func_extended: add three api tests
The first two are heavily reused version of the gl-3.0 API tests from Ian, just
adapted to test the two APIs for GL_ARB_blend_func_extended,

the third is a simple test that glBlendFunc can accept the new tokens, and
the changed behaviour of the GL_SRC_ALPHA_SATURATE token.

tested on fglrx which fails two of the tests because it returns the wrong
error.

Signed-off-by: Dave Airlie <airlied@linux.ie>