glsl: Add tests for glGetActiveAttrib
Verify that glGetActiveAttrib and GL_ACTIVE_ATTRIBUTES return sensible
data under a variety of circumstances.  There are checks for all the
types and built-in attributes in GLSL 1.10, 1.20, and 1.30.

All of the subcases pass on NVIDIA's closed-source Linux driver and
AMD's closed-source Linux driver.

These test fails horribly on Mesa 7.11.  Matrix attributes are
reported multiple times (once per column of the matrix) and built-in
attributes are never reported.  Mesa master fixes the matrix issues,
but it still has problems with built-in attributes.  There is a patch
on the mesa-dev mailing list to fix that issue.

v2: Terminate the list of attributes with a NULL name instead of
having an explicit count (suggested by Eric Anholt).  Add missing test
for ftransform.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=43138
Reviewed-by: Eric Anholt <eric@anholt.net>