glx_ext_import_context: Verify function of glXFreeContextEXT
This test fails on NVIDIA's closed-source driver for a couple reasons.
First, inspite of what the spec says, they appear to release the
server-side context when glXFreeContextEXT is called.  Second, they do
not generate a GLXBadContext error when a NULL context is freed.

This test fails on AMD's closed-source driver because it hits an
assertion failure in XCB:

xcb_io.c:515: _XReply: Assertion `!dpy->xcb->reply_data' failed.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Brian Paul <brianp@vmware.com>