Add a test of alignment of transform feedback outputs.
Certain graphics hardware (notably Intel i965) has restrictions on
unaligned memory accesses.  However, transform feedback can be
configured in such a way that data is not aligned on a natural
boundary (e.g. a vec4 is 16 bytes wide, but is only required to be
aligned to a multiple of 4 bytes).  This test verifies that transform
feedback data is written to memory correctly regardless of how the
data is aligned.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>