Fix infinite loop in piglit_is_extension_in_string.
If piglit_is_extension_in_string() finds an extension whose name
begins with the name it is searching for, but is not equal to the name
it is searching for (e.g. it is searching for EXT_transform_feedback
but it finds EXT_transform_feedback2), we need it to advance to the
end of the substring it's found before restarting the search,
otherwise it will loop indefinitely.

Prevents a hang when running some transform feedback tests on nVidia's
proprietary Linux driver.

Reviewed-by: Brian Paul <brianp@vmware.com>