Transform feedback: immediately reuse XFB output as vertex input.
This test verifies that transform feedback data is properly handled if
it is immediately re-used as vertex input, with as little change as
possible to GL state in between.  This verifies that the
implementation maintains proper cache coherency between transform
feedback outputs and vertex inputs.

Verified on the nVidia proprietary Linux driver.

Reviewed-by: Eric Anholt <eric@anholt.net>