Test transform feedback with intervening glReadPixels.
This test verifies that transform feedback continues to work correctly
if glReadPixels is executed while it is in progress.

This is an important case to test on Mesa's i965 driver, because
glReadPixels forces the driver to start a new batch buffer, and at the
beginning of each batch buffer the driver needs to update the
transform feedback buffer pointers (in case they were used by another
process in between the two batches).

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Eric Anholt <eric@anholt.net>