Transform feedback: test edge cases in overflow checking.
This test checks for off-by-one errors in the mechanism to prevent
overflow of the transform feedback buffer, by testing a wide variety
of buffer sizes, numbers of varying components, numbers of primitives
drawn, and primitive modes.  In each case it verifies that the data
written to the transform feedback buffer is appropriately truncated,
and that the counters GL_TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN and
GL_PRIMITIVES_GENERATED are set correctly.

Verified on the nVidia proprietary Linux driver, and on Mesa i965.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>