Modify piglit_check_gl_error() to return a boolean.
Previously piglit_check_gl_error() would terminate the test if an
error occurred.  However, there are many cases when we would prefer to
simply report the error, and continue testing.

Where it was convenient, I've replaced calls of the form

piglit_check_gl_error(..., PIGLIT_FAIL);

with

pass = piglit_check_gl_error(...) && pass;

So that the test can continue executing.

In other cases, I replaced it with

if (!piglit_check_gl_error(...))
        piglit_report_result(PIGLIT_FAIL);

Which is equivalent to the old functionality.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>