fbo-blit-d24s8: don't try to blit betweendifferent Z/stencil depths
The docs for glBlitFramebuffer say GL_INVALID_OPERATION is generated
if the source/dest depth/stencil buffer formats do no match.  The spec
is a bit vague on what's meant by formats (do just the component depths
need to match, or must the hw formats match?) but clearly if the depths
don't match, the formats can't match.

This fixes a bogus error that's generated when testing swrast which
usually uses a 16-bit depth buffer for the window and a 24-bit Z
renderbuffer for the FBO.

v2: just report PIGLIT_SKIP if the bit depths don't match (per Eric)

Reviewed-by: Eric Anholt <eric@anholt.net>