Transform feedback: Update builtin-varyings to test size and type.
This patch expands the "EXT_transform_feedback/builtin-varyings" tests
so that in addition to verifying that the transform feedback data is
written correctly, they verify that GetTransformFeedbackVarying()
returns the correct size and type values for each builtin.

This is particularly important to test for gl_ClipDistance, since some
drivers (e.g. Mesa i965) represent it internally as an array of 1 or 2
vec4's, but GetTransformFeedbackVarying() needs to report it as an
array of 1 to 8 floats.

Reviewed-by: Eric Anholt <eric@anholt.net>