Stop using glewIsSupported
Several tests were using the function glewIsSupported() to check for
extensions and GL versions.  This patch changes the tests to use
piglit_get_gl_version() and piglit_is_extension_supported(), which
provide the same functionality in a way that is not dependent on GLEW.

Reviewed-by: Eric Anholt <eric@anholt.net>
Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>

v2: Follow piglit convention of including "GL_" in the string passed
to piglit_is_extension_supported().