piglit-run: support storing results outside of piglit_dir
piglit-run.py takes the directory to store the results in as a command
line argument, yet always chdir's to piglit_dir first before accessing it.
When running piglit from outside of piglit_dir, this is surprising to the
user and will fail when piglit is packaged in a distribution and the user
can't write to piglit_dir.

This came up while packaging piglit for Fedora:
https://bugzilla.redhat.com/show_bug.cgi?id=773021
where all piglit code is placed under /usr/lib64/piglit/ and symbolic
links to the piglit-* scripts are created in /usr/bin/.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>
Signed-off-by: Scott Tsai <scottt.tw@gmail.com>