polygon-mode: replace memcmp() with better pixel probing
Previously, we compared the reference image and the test image with
memcmp().  That failed when the driver used different drawing paths
for glPolygonMode() vs. regular filled/line/point drawing.

The new test examines the rendering with 3x3-pixel probes to determine if
a polygon was drawn filled, outlined or as points.

Reviewed-by: José Fonseca <jfonseca@vmware.com>