Script to produce results in a format that Jenkins-CI understand.
I've been using piglit on Jenkins-CI (aka Hudson-CI) for a long time, but
it did not use piglit framework, and implied manually duplicating test
lists.  It was an easy way to get things started, but far from ideal.

This is a first step of an attempt to properly add Jenkins-CI support
to piglit, so that new upstream piglit tests can be picked up automatically
without manual intervention.