Modify shader_runner to skip tests with unsigned uniforms when GL<3.0
On implementations that support GLSL version 1.30 but not GL version
3.0 (for example Mesa, when built without floating point texture
support), it is impossible to test unsigned uniforms, because the
Uniform*ui{v} functions are not available.  So any shader_runner tests
that use unsigned uniforms should be skipped.

Since it's rare for implementations to support GLSL version 1.30 but
not GL version 3.0, we don't want to require test writers to remember
to add "GL>=3.0" to the top of their test scripts.  So instead, this
patch modifies shader_runner to automatically skip any test that uses
unsigned uniforms if the GL version is less than 3.0.

Avoids 404 bogus "crash" results in
spec/glsl-1.30/execution/built-in-functions when testing with a
version of Mesa that was built without floating point texture support.

Reviewed-by: Eric Anholt <eric@anholt.net>
Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>