framework: Fix filtering of excluded tests.
del dict[key] will raise a KeyError if the key isn't actually in the
dictionary.  This can happen if, say, a test didn't match the -t option
and subsequently matched a -x option.

Also, deleting dictionary items while iterating over them is apparently
not safe in Python 3, so we may as well just create a new dictionary
with only the entries we want.

This is simpler anyway.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>