Fix ARB_draw_instanced feature-macro-disabled tests.
If the GL supports the GL_ARB_draw_instanced extension, the preprocessor
macro must exist regardless of whether or not the shader enables the
extension.  (In fact, this allows them to check if it's supported!)

This patch changes the tests to only run if GL_ARB_draw_instanced is
_not_ supported by the driver, in which case it's an error to define
the preprocessor macro.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>