html: Use the vastly faster "fixed" table layout algorithm.
The usual "auto" table layout algorithm makes the browser look at the
contents of each row in the table in order to determine the layout.  We
have 10,000 rows.  This is insanely expensive.

The "fixed" algorithm just uses the <col/> info, and possibly the first
row.  This works out just fine.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Dylan Baker <baker.dylan.c@gmail.com>