Stop using GLEW extension macros
Several tests were checking for the presence of GL extensions using
GLEW macros.  This patch changes the tests to use
piglit_is_extension_supported(), which provides the same functionality
in a way that is not dependent on GLEW.

v2: Follow piglit convention of including "GL_" in the string passed
to piglit_is_extension_supported().

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>
Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>