Add a call to glewInit() to windowoverlap.c
Previously this test didn't initialize GLEW, presumably because it
only uses GL 1.0 functions, and those functions bypass GLEW.  But once
we switch to using piglit-dispatch instead of GLEW, we will need to
initialize it in order for any GL functions to work.  This patch adds
a call to glewInit() in the place where a call to
piglit_dispatch_default_init() will eventually be needed.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>