glean: use piglit-util and initialize GLEW.
Later patches in this series will introduce a new mechanism for
binding Piglit to GL functions, called piglit-dispatch.  This
mechanism will reside in the piglit-util library, and will require
initialization.

To ease the transition, this patch causes Glean to link to
piglit-util, and initializes GLEW using the glewInit() function.  When
GLEW is replaced with piglit-dispatch, the call to glewInit() will
serve to initialize new piglit-dispatch mechanism.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>