Add glewInit() calls to GLX_ARB_create_context tests.
Previously these tests didn't initialize GLEW, because they only use
GLX functions and functions defined in GL 1.0 and 1.1 (all of which
bypass GLEW).  But once we switch to using piglit-dispatch instead of
GLEW, we will need to initialize it in order for any GL functions to
work.  This patch adds a call to glewInit() in the place where a call
to piglit_dispatch_default_init() will eventually be needed.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>