Use nearest filtering in the texelFetch() and textureSize() tests.
According to the OpenGL 3.0 specification, page 214, textures with
internal integer format must use NEAREST for the mag filter and either
NEAREST or NEAREST_MIPMAP_NEAREST for the min filter.  Otherwise, the
texture should be incomplete.

Both texelFetch() and textureSize() ignores the filter modes, so there's
no harm in switching to nearest.  Obviously we need the texture to be
complete.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Eric Anholt <eric@anholt.net>