Add a test for GL_EXT_unpack_subimage on GLES2
The GL_EXT_unpack_subimage extension just allows GL_UNPACK_ROW_LENGTH,
GL_UNPACK_SKIP_PIXELS and GL_UNPACK_SKIP_ROWS on GLES2. This tries
setting those three values and if the extension is available it will
assert that no error is thrown, otherwise it will assert that
GL_INVALID_ENUM is thrown. It will then try to upload some texture
data with those settings in such a way that if the values were
correctly set it will have one pair of pixels otherwise it will have
another. The texture is then rendered to the framebuffer and the
pixels are compared so it can verify whether the unpack settings were
handled correctly.

As far as I can tell, this is the first GLES2 test so the patch also
creates an all_es2.tests file to contain it.