shader_runner: Rename generic "mode_table" variable for clarity.
OpenGL has a lot of different kind of modes, so "mode_table" is a bit
too generic of a name.  Rename it to "drawing_mode_table" for clarity.

While we're at it, unify the "mode_table" struct with the
"string_to_enum" struct in handle_texparameter, as they're identical,
and the latter has a more descriptive name.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>