shader_runner: Add support for arbitrary texparameter targets.
Previously, we hardcoded the sampler name into the texparameter command
itself (i.e. texparameter2DArray) and relied on the main sscanf loop to
parse that for us.

That doesn't work well once you start adding more than a handful of
texture tests.  Parsing the texture target in handle_texparameter()
allows us to cleanly support the full set.

This preserves the spaceless command names for compatibility (and ease
of review).  The next commit will fix that.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>