Add a test of the accuracy of MSAA rendering.
This is the test I've been using to prototype MSAA on the i965
driver.  Unlike the existing EXT_framebuffer_multisample
tests (which so far mostly check corner cases at the API level),
this test verifies that correct pixels are rendered when MSAA is
enabled.  It's a bit long for a typical piglit test, since in
order to verify that the MSAA color, depth, and stencil buffers
are all configured correctly, it has to render some fairly
complex scenes.