glut_waffle: Add libglut_waffle [v3]
This patch adds libglut_waffle to the build. No tests use it yet.

glut_waffle is an implementation of the subset of GLUT used by Piglit.

glut_waffle is transitionary only and not intended to be a permanent
addition to Piglit. Its purpose is to make Piglit's transition from GLUT
to Waffle go smoothly. Once the transition is complete, piglit-framework.c
will be updated to use Waffle directly, and libglut_waffle will be
removed.

v2: Remove else blocks in glutChooseConfig that just set default values,
    per Anholt.
v3: Fix typo that set WAFFLE_ACCUM_BUFFER to false.

CC: Eric Anholt <eric@anholt.net>
Reviewed-by: Pauli Nieminen <pauli.nieminen@intel.com> [v1]
Reviewed-by: Jordan Justen <jordan.l.justen@intel.com> [v1]
Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com> [v1]
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>