Test rendering to various miplevels of depth and stencil textures.
This patch adds a new test, "depthstencil-render-miplevels", which
verifies that we can render to all miplevels of depth and stencil
textures.  It exposes two bugs in Mesa's current i965 driver:

- Miptree offsets aren't being calculated correctly for HiZ and
  separate stencil buffers, causing miplevels to overlap in those
  buffers.

- When binding the same texture to the GL_DEPTH_ATTACHMENT and
  GL_STENCIL_ATTACHMENT points, Mesa is trying to share internal data
  structures even if the miplevels don't match; as a result,
  transitioning from one miplevel to another doesn't work.

v2: Added comment "/* Anonymous namespace */" on the final closing
brace.  Adjusted the naming scheme for the subtests to allow
additional format combinations to be added in the future.  Also added
format combinations involving z32f and z32f_s8.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>