glsl-1.10: Add a test for a bad shader not breaking a linked program.
This was a test I came up with from a description of a failure in the
webgl conformance suite, but which I don't know the actual sequence of
GL calls for.  It passes on Mesa, so it didn't catch what I was hoping
for.  I don't see previous coverage of this in piglit, though.