framework: Split a TestProfile.prepare_test_list helper out of run().
Before scheduling or running the tests, run() prepared the final list of
tests to run: first, flatten the Group() hierarchy; second, filter out
tests based on the -t and -x options.

It makes sense to have this as a helper function.  Doing so will also
enable other utilities that (for example) print a list of tests that
would be run and their command line programs.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Eric Anholt <eric@anholt.net>