junit: Tweak the test hierarchy mapping.
The junit format used by Jenkins came from Java land, and has three
levels: packages, classes, and test methods; so mapping a generic test
hierarchy to these is far from straightforward.

This change tries to rearrange the mapping, in such a way that the test
results as presented by Jenkins look better and are easier to navigate.