copyteximage: fix the test for legal texture formats
The code for checking if the texture format is legal was missing some
cases (such as GL_RGB16F).  We also need to check if the format is
legal for the texture target.  Compressed textures can't be used with
GL_TEXTURE_1D for example.

Refactor the format checking code to make it easier to read and expand
in the future.

Fixes https://bugs.freedesktop.org/show_bug.cgi?id=49124

Reviewed by: Anuj Phogat <anuj.phogat@gmail.com>