Test that MSAA blits work properly even when unaligned.
This test verifies that blits involving MSAA buffers work properly
even if the source and destination rectangles are located in different
parts of their respective framebuffers, and even if the source and
destination rectangles are not aligned to powers of two.

These are necessary corner cases to check in i965, where blits that
involve MSAA buffers frequently require special handling when the
source and destination rectangles are different, or are not nicely
aligned.

v2: Use a tile size of 49 instead of 50, so that source and
destination rectangles are even more poorly aligned.  This exposed a
bug in i965/gen6.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>