Add test to draw multisample lines with GL_LINE_SMOOTH enabled
This test varifies that GL_LINE_SMOOTH is ignored in case of multisample
lines.

Note: Smooth lines and multisampling seems to work fine separately but this
test fails on AMD catalyst drivers. Test passes on NVIDIA drivers.

V2: Line smoothing is applied uniformly to all attached buffers. So, It is
    redundant to test on depth and stencil buffers.

Signed-off-by: Anuj Phogat <anuj.phogat@gmail.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>