glsl-1.20 / glsl-1.30: Add tests for uniforms with initializers
These first tests simply set the value of an uniform with an
initializer in the shader source and verifies that the uniform has
that value when the shader is executed.

v2: Use '{0}{1}' formatting instead of '{}{}'.  The latter is added in
Python 2.7, but piglit only requires Python 2.6.  This was pointed out
by Jordan Justen.

v3: Clean up the generation of the if-statement based on feedback from
Paul Berry.  Also clean up the list types used to generate tests.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>