glsl-1.20 / glsl-1.30: Add tests for uniforms with initializers from a constant
These tests are similar to the previous tests except the uniforms are
initialized using a variable declared as const.  Since any
compile-time constant expression can be used to initialize a uniform,
this should also work.

v2: Use '{0}{1}' formatting instead of '{}{}'.  The latter is added in
Python 2.7, but piglit only requires Python 2.6.  This was pointed out
by Jordan Justen.

v3: Clean up the generation of the if-statement based on feedback from
Paul Berry.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>