glsl-1.20 / glsl-1.30: Add tests for uniforms with initializers modified by GL API
In these tests the uniforms are initialized with one set of values in
the shader source, but the values are changed using the GL API.  The
shaders expect the values set by the GL API.

v2: Use '{0}{1}' formatting instead of '{}{}'.  The latter is added in
Python 2.7, but piglit only requires Python 2.6.  This was pointed out
by Jordan Justen.

v3: Clean up the generation of the if-statement based on feedback from
Paul Berry.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>