msaa: Make a new Fbo::try_setup() function.
In contrast to the Fbo::setup() function (which terminates the test if
the framebuffer is incomplete), Fbo::try_setup() merely returns a
boolean indicating whether the framebuffer is complete.  This will be
necessary to allow different color formats to be tested, since we
won't want to terminate the test if a non-renderable color format is
attempted.

Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>