msaa: Test that MSAA works properly for all core FBO color formats.
This test builds on the existing FBO and MSAA testing infrastructure
to verify proper functioning of MSAA for various formats.

All core formats which are framebufferbuffer complete are tested.  All
formats which are not framebuffer complete are skipped.

Tests for non-core and depth/stencil formats will be added in later
patches.

Verified using the proprietary nVidia driver for Linux--all formats
pass except the LUMINANCE_ALPHA formats, which fail due to a driver
bug (when rendering to these formats, the nVidia driver maps the red
channel to both luminance and alpha, instead of mapping red to
luminance and alpha to alpha).

Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>