cmake: If USE_WAFFLE, disable tests that use quirky GLUT functions [v2]
If USE_WAFFLE is enabled, do not build the tests below.  The tests use
quirky use GLUT functions that will not be implemented in Waffle.
    fbo-depth-sample-compare: Uses glutSolidSphere.
    windowoverlap: Uses glutCreateSubWindow.
    texwrap: Uses glutBitmapCharacter.
    texture-pakced-formats: Uses glutBitmapCharacter.

Eventually, these tests need to be repaired so that they no longer require
the quirky GLUT functions.  However, the Waffle integration shouldn't be
blocked by this small set of tests. Work on reparing these tests can
proceed in tandem with the Waffle integration.

v2: Fix indentation.

Reviewed-by: Pauli Nieminen <pauli.nieminen@intel.com>
Reviewed-by: Jordan Justen <jordan.l.justen@intel.com>
Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>