Reproduce a constant propagation bug in Mesa (indexing into vectors).
As of 5/19/12, Mesa's constant propagation optimization pass fails to
properly account for the behaviour of array indexing applied to
vectors.  If v is a vector and i is an integer, it assumes that

v[i] = ...;

kills the x component of v.  This is not necessarily the case--it
might kill any component of v, depending on the value of i.

Reviewed-by: Jose Fonseca <jfonseca@vmware.com>