util: Fork piglit-framework.c
Copy piglit-framework.c to piglit-framework-fbo.c. The following commits
will refactor each so that all the piglit_framework_fbo code is contained
in piglit-framework-fbo.c. This copy is done before the refactor so that
git merges go smoothly.

The piglit_framework_fbo code needs to be relocated to its own file
because following commits will add an implementation atop Waffle.
Currently, the piglit_framework_fbo code is only implemented for GLX. The
increase in code size and #ifdef complexity warrants moving it to its own
file.

Reviewed-by: Pauli Nieminen <pauli.nieminen@intel.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>