cmake: Stop GLES tests from linking to glut_egl
Many CMakeLists.gles{1,2}.txt explicitly linked to glut_egl. But all GLES
tests should now use glut_waffle.

This patch kills the directives to link to glut_egl, but does not replace
them with a directive to link to glut_waffle. It is not needed, because
`link_libraries(glut_waffle)` exists in the global GLES makefiles,
$PIGLIT_TOP/cmake/target_api/gles{1,2}/CMakeLists.txt.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>