fbo: Add base internal format info to fbo-formats.h
This patch augments the format_desc structure (used by fbo-formats.h
to describe a format to be tested) with a base_internal_format field,
which indicates the base internal format corresponding to each
particular internal format (e.g. the base internal format of
GL_R3_G3_B2 is GL_RGB).

The field test_desc::base (which carries the same information, but
only for depth/stencil formats, and is sometimes incorrect) will be
removed in a later patch.

Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>