fbo: Rearrange if tests for GL_DEPTH32F_STENCIL8 behaviour.
The tests fbo-clear-formats and fbo-generatemipmap-formats need
special behaviour for GL_DEPTH32F_STENCIL8.  This patch rearranges the
if statements that implement that special behaviour so that they no
longer rely on the fact that GL_DEPTH32F_STENCIL8 is (incorrectly)
categorized as having a base internal format of GL_DEPTH_COMPONENT
rather than GL_DEPTH_STENCIL.  This will allow the incorrect
categorization to be removed in a future patch.

Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>