msaa: Make it possible to reconfigure the fbo after initial use.
This patch makes it safe to re-initialize the state of an Fbo object
to a new configuration.  This will be necessary to allow this object
to be used for testing formats, since one piglit test will have to
test several different Fbo formats in sequence.

To reflect the fact that initialization of the Fbo object is no longer
a one-time action, we now use the name setup() for the function that
sets the Fbo state, rather than init().

Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>