msaa: Make it possible to change output type when drawing sunbursts.
In order to test that MSAA works properly for integer framebuffers, we
will need to be able to output ivec4 and uvec4 types when rendering
the test image.  This patch makes that possible by adapting the
Sunburst GLSL program so that it can be compiled to output ivec4's,
uvec4's, or vec4's.

Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>