util: Remove and refactor weak global symbols in libpiglitutil
This patch's goal is to change the the behavior of the dataflow between
test executables and libpiglitutil. To accomplish that, it removes some
weak global symbols in libpiglitutil and changes the semantics of others.
All changes are located in the piglit-framework-* files.

(I believe that, after this patch, libpiglit can be built as a shared
library on MSVC because it no longer defines weak symbols. But I haven't
tested this.)

Before this patch, the dataflow operated solely via global symbols. Some
symbols were used by the test executable to provide data to libpiglitutil;
some symbols were used for the other direction; and, confusingly, some
symbols provided bidirectional dataflow and conveyed different data
according to direction. The dataflow worked like this:

    Direction A: test executable => libpiglitutil
        - Test behavior, via global functions piglit_init and
          piglit_display.
        - Initialization parameters, via global variables piglit_width,
          piglit_height, and piglit_window_mode.

    Direction B: libpiglitutil => test executable
        - User-requested test options, via global variables piglit_use_fbo
          and piglit_automatic.
        - Dynamic test state, via global variables piglit_width and
          piglit_height.

Observe that, in direction A, piglit_width and piglit_height serve as
initialization parameters. In direction B, they serve to convey dynamic
state.

After this patch, global symbols are used for dataflow in only direction
B, and that dataflow remains unchanged. The dataflow for direction A now
works like this:

    Direction A: test executable => libpiglitutil
        - Test behavior, via function pointers piglit_gl_test_info::{init,display}.
        - Initialization parameters, via piglit_gl_test_info::window_{width,height,visual}.

As a consequence of this change, the following symbols in libpiglitutil
are no longer used and are removed:
    - piglit_init
    - piglit_display
    - piglit_window_mode

Also, since piglit_width and piglit_height are no longer defined by test
executables, but are defined exclusively by libpiglitutil, they need no
longer be declared as weak symbols.

Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>