Fix gen_uniform_initializer_tests.py for out-of-tree builds.
Previously, the script opened template files relative to the current
working directory.  This failed for out-of-tree builds because the
script is run from within the build tree, not the source tree.

This patch fixes the problem by using Python's built-in __file__
variable to determine the path of the source file, and then looking
for the template files relative to it.

Signed-off-by: Paul Berry <stereotype441@gmail.com>