security/initialized-texmemory - test contents of unitialized texture memory
This is the first of a new category of piglit tests that exercise
things related to security.

The first test, initialized-texmemory, creates a texture image with
unspecified contents (pixels=NULL).  We draw a polygon with the texture
and see what we get.  Ideally, we'd get a constant/black image and not
some random VRAM data from another application or another user (like a
window image).  The OpenGL spec doesn't say that texture memory should
be initialized in this situation so if the test fails we return "WARN"
rather than "FAIL".

Reviewed-by: José Fonseca <jfonseca@vmware.com>